pyproject.toml
[project]
name = "model-doc-automation"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.0.2",
    "langchain-core>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langgraph-checkpoint>=3.0.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "scipy>=1.16.2",
    "tabulate>=0.9.0",
]



[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

